moved components into flask directory and updated imports
