+++ ui/pages/settings/flask/snaps-list-tab/index.scss
