+++ ui/pages/settings/flask/snaps-list-tab/snap-list-tab.stories.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Provider } from 'react-redux';
-import configureStore from '../../../store/store';
-import testData from '../../../../.storybook/test-data';
+import configureStore from '../../../../store/store';
+import testData from '../../../../../.storybook/test-data';
 import SnapListTab from './snap-list-tab.component';
 
 // Using Test Data For Redux
