+++ ui/pages/settings/flask/snaps-list-tab/view-snap/index.js
