+++ ui/pages/settings/flask/snaps-list-tab/view-snap/view-snap.component.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from '../../../../components/ui/button';
-import Typography from '../../../../components/ui/typography';
-import { useI18nContext } from '../../../../hooks/useI18nContext';
+import Button from '../../../../../components/ui/button';
+import Typography from '../../../../../components/ui/typography';
+import { useI18nContext } from '../../../../../hooks/useI18nContext';
 import {
   TYPOGRAPHY,
   COLORS,
-} from '../../../../helpers/constants/design-system';
-import SnapsAuthorshipPill from '../../../../components/app/flask/snaps-authorship-pill';
-import Box from '../../../../components/ui/box';
-import ToggleButton from '../../../../components/ui/toggle-button';
-import PermissionsConnectPermissionList from '../../../../components/app/permissions-connect-permission-list/permissions-connect-permission-list';
+} from '../../../../../helpers/constants/design-system';
+import SnapsAuthorshipPill from '../../../../../components/app/flask/snaps-authorship-pill';
+import Box from '../../../../../components/ui/box';
+import ToggleButton from '../../../../../components/ui/toggle-button';
+import PermissionsConnectPermissionList from '../../../../../components/app/permissions-connect-permission-list/permissions-connect-permission-list';
 
 function ViewSnap({ snap, onRemove, onToggle }) {
   const t = useI18nContext();
