+++ ui/pages/settings/flask/snaps-list-tab/view-snap/view-snap.container.js
