+++ ui/pages/settings/settings.component.js
@@ -28,7 +28,7 @@ import InfoTab from './info-tab';
 import SecurityTab from './security-tab';
 import ContactListTab from './contact-list-tab';
 import ExperimentalTab from './experimental-tab';
-import SnapListTab from './snaps-list-tab';
+import SnapListTab from './flask/snaps-list-tab';
 
 class SettingsPage extends PureComponent {
   static propTypes = {
